// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/abwab-tree-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AbwabTreeDoorDto } from '../models/abwab-tree-door-dto';
import { AbwabTreeSectionDto } from '../models/abwab-tree-section-dto';
export interface AbwabTreeDto {
  doors: Array<AbwabTreeDoorDto>;
  sections: Array<AbwabTreeSectionDto>;
  version: string | null;
}